[S13P31A406-157] 밈 상세 조회 api 구현

See merge request s13-final/S13P31A406!93

+++ backend/src/main/java/com/a404/duckonback/controller/MemeController.java
@@ -111,6 +111,8 @@ public class MemeController {
             @AuthenticationPrincipal CustomUserPrincipal principal,
             @PathVariable Long memeId
     ) {
+        if(principal == null) throw new CustomException("로그인이 필요합니다.", ErrorCode.USER_NOT_AUTHENTICATED);
+
         memeService.createFavorite(principal.getId(), memeId);
         return ResponseEntity.ok(ApiResponseDTO.success(SuccessCode.MEME_FAVORITE_CREATED, null));
     }
@@ -121,6 +123,8 @@ public class MemeController {
             @AuthenticationPrincipal CustomUserPrincipal principal,
             @PathVariable Long memeId
     ) {
+        if(principal == null) throw new CustomException("로그인이 필요합니다.", ErrorCode.USER_NOT_AUTHENTICATED);
+
         memeService.deleteFavorite(principal.getId(), memeId);
         return ResponseEntity.ok(ApiResponseDTO.success(SuccessCode.MEME_FAVORITE_DELETED, null));
     }
@@ -162,6 +166,18 @@ public class MemeController {
         return ResponseEntity.ok(
                 ApiResponseDTO.success(SuccessCode.MEME_TOP10_RETRIEVE_SUCCESS, res)
         );
+    }
+
+    @Operation(
+            summary = "밈 상세 조회",
+            description = "밈 기본정보, 태그 이름 리스트, 생성자 정보를 반환합니다."
+    )
+    @GetMapping("/{memeId}/detail")
+    public ResponseEntity<ApiResponseDTO<MemeDetailDTO>> getMemeDetail(
+            @PathVariable Long memeId
+    ) {
+        MemeDetailDTO detail = memeService.getMemeDetail(memeId);
+        return ResponseEntity.ok(ApiResponseDTO.success(SuccessCode.MEME_RETRIEVE_SUCCESS, detail));
     }
 
     @Operation(
